Fix numpy v1.10 issue in metutils.convert()

+++ Utilities/metutils.py
@@ -385,7 +385,7 @@ def convert(value, inunits, outunits):
 
     """
     startValue = value
-    value = np.array(value)
+    value = np.array(value, dtype=float)
     if inunits == outunits:
         # Do nothing:
         return value
